move files to explore dir
